+++ frontend/src/components/index.js
@@ -1,5 +1,5 @@
 export { default as TopBar } from "./TopBars/TopBar";
 export { default as Button } from "./Button/Button";
-export { default as Logo } from "./logo/Logo";
+export { default as Logo } from "./logoo/Logo";
 export { default as Cards } from "./Cards/Cards";
 export { default as Card } from "./Card/Card";
