+++ frontend/src/components/logoo/img/fruit.png
