+++ frontend/src/components/logoo/img/fruit_original.png
